// src/app/layout/manage/model/request-management/management-permission.model.ts
export class ManagementPermission {
    nid_permission: number = 0;
    nid_project: number = 0;
    nid_request: number = 0;
    nidresource: number = 0;
    dstart: string = "";
    dfinish: string = "";
    dregister: string = "";
    userregister: number = 0;
    dupdate: string = "";
    userupdate: number = 0;
}